Move Http namespace under Logger
