+++ tests/phpunit/Logger/BadgeLoggerTest.php
@@ -37,8 +37,8 @@ namespace Infection\Tests\Logger;
 
 use Infection\Environment\BuildContextResolver;
 use Infection\Environment\StrykerApiKeyResolver;
-use Infection\Http\StrykerDashboardClient;
 use Infection\Logger\BadgeLogger;
+use Infection\Logger\Http\StrykerDashboardClient;
 use Infection\Metrics\MetricsCalculator;
 use Infection\Tests\Double\OndraM\CiDetector\ConfigurableEnv;
 use Infection\Tests\EnvVariableManipulation\BacksUpEnvironmentVariables;
